+++ src/Config/SQLServerDriverConfig.php
@@ -32,7 +32,7 @@ class SQLServerDriverConfig extends DriverConfig
         bool $queryCache = true,
         bool $readonlySchema = false,
         bool $readonly = false,
-        bool $datetimeWithMicroseconds = false,
+        array $options = [],
     ) {
         /** @psalm-suppress ArgumentTypeCoercion */
         parent::__construct(
@@ -43,7 +43,7 @@ class SQLServerDriverConfig extends DriverConfig
             queryCache: $queryCache,
             readonlySchema: $readonlySchema,
             readonly: $readonly,
-            datetimeWithMicroseconds: $datetimeWithMicroseconds,
+            options: $options,
         );
     }
 }