+++ src/Config/SQLiteDriverConfig.php
@@ -33,7 +33,7 @@ class SQLiteDriverConfig extends DriverConfig
         bool $queryCache = true,
         bool $readonlySchema = false,
         bool $readonly = false,
-        bool $datetimeWithMicroseconds = false,
+        array $options = [],
     ) {
         /** @psalm-suppress ArgumentTypeCoercion */
         parent::__construct(
@@ -44,7 +44,7 @@ class SQLiteDriverConfig extends DriverConfig
             queryCache: $queryCache,
             readonlySchema: $readonlySchema,
             readonly: $readonly,
-            datetimeWithMicroseconds: $datetimeWithMicroseconds,
+            options: $options,
         );
     }
 }