+++ CHANGELOG.md
@@ -4,8 +4,9 @@ v2.3.0 (in progress)
 -------------------
 - Add supporting for array values in the `IN` and `NOT IN` operators by @roxblnfk (#69, #70, #71)
 - Add `PdoInterface` as possible return type of the `Driver::getPDO()` method by @roxblnfk (#76)
-- Add parameter `datetimeWithMicroseconds` to all driver configs to store the DateTime string with microseconds in the 
-  database by @msmakouz (#86)
+- Add array parameter `options` to all driver configs to pass additional driver options.
+  To store a DateTime with microseconds, you can pass the `datetimeWithMicroseconds` with value **true**
+  in the new options parameter by @msmakouz (#86)
 - Add ability to define `size` in the `datetime` column by @msmakouz (#86)
 
 v2.2.2 (27.09.2022)